pyproject.toml
[tool.poetry]
name = "graphbench-kg"
version = "0.1.0"
description = "Benchmark framework for graph-based RAG pipelines on multi-hop QA"
authors = ["Rohan Jain <rohanjain2312@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Rohanjain2312/graphbench"
repository = "https://github.com/Rohanjain2312/graphbench"
documentation = "https://github.com/Rohanjain2312/graphbench/blob/main/docs/index.md"
keywords = ["knowledge-graph", "RAG", "GNN", "benchmark", "question-answering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{ include = "graphbench" }]

[tool.poetry.dependencies]
python = "^3.10"

# Graph DB
neo4j = "^5.18.0"

# Vector search
faiss-cpu = "^1.7.4"

# Embeddings
sentence-transformers = "^2.7.0"

# ML / GNN
torch = "^2.2.0"
torch-geometric = "^2.5.0"

# Data
datasets = "^2.19.0"
pandas = "^2.2.0"
numpy = "^1.26.0"

# Config
pydantic-settings = "^2.2.0"
python-dotenv = "^1.0.0"

# Experiment tracking
wandb = "^0.17.0"

# NLP utilities
transformers = "^4.40.0"
accelerate = "^0.30.0"
bitsandbytes = "^0.43.0"

# Community detection
python-louvain = "^0.16"
networkx = "^3.3"

# Demo
gradio = "^4.31.0"
pyvis = "^0.3.2"
plotly = "^5.22.0"

# Utilities
tqdm = "^4.66.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.14.0"
black = "^24.4.0"
isort = "^5.13.0"
ruff = "^0.4.0"
pre-commit = "^3.7.0"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short -m 'not slow'"
markers = [
    "slow: marks tests that require model downloads or GPU (skip with -m 'not slow')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
